requirements.txt
ultralytics
pathlib
argparse
PyYAML